Boot rom should be accessed as uint16_t, not as uint8_t

+++ src/hagemu_core/boot.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-uint8_t boot_read(uint8_t address);
+uint8_t boot_read(uint16_t address);
 
 #endif
